+ Adjusting some stuff * Added RoundedStrokeButton * Moved RoundedButton into button/ directory * Added ic_flash_off and use it as default state
