+++ app/src/main/res/drawable/bg_rounded_ripple_stroke.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ripple
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        android:color="?android:attr/colorControlHighlight">
+    <item>
+        <shape
+                android:shape="rectangle">
+            <solid android:color="@color/white"/>
+            <corners android:radius="12dp"/>
+            <stroke android:width="2dp" android:color="?attr/colorPrimary"/>
+        </shape>
+    </item>
+</ripple>
