+++ app/src/main/res/layout/activity_camera.xml
@@ -83,7 +83,7 @@
             android:layout_width="40dp"
             android:layout_height="0dp"
             android:layout_marginTop="24dp"
-            android:src="@drawable/ic_flash"
+            android:src="@drawable/ic_flash_off"
             android:layout_marginEnd="24dp"
             android:contentDescription="@string/torch_camera"
             android:elevation="16dp"
