+++ app/src/main/res/layout/activity_login.xml
@@ -86,7 +86,7 @@
             app:layout_constraintBottom_toTopOf="@id/btn_login"
             android:id="@+id/tv_forgot_password"/>
 
-    <io.github.skincanorg.skincan.widget.RoundedButton
+    <io.github.skincanorg.skincan.widget.button.RoundedButton
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:text="@string/login"
