+++ app/src/main/res/layout/activity_scanner.xml
@@ -10,26 +10,26 @@
 
     <androidx.cardview.widget.CardView
             android:layout_width="match_parent"
-            android:layout_height="0dp"
+            android:layout_height="wrap_content"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintDimensionRatio="9:16"
             app:cardCornerRadius="8dp"
             app:cardElevation="0dp"
             android:id="@+id/image_preview_wrapper">
 
         <ImageView
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
+                android:layout_height="wrap_content"
                 android:contentDescription="@string/captured_image_preview"
+                tools:src="@drawable/ic_picture"
                 android:id="@+id/iv_image_preview"/>
 
     </androidx.cardview.widget.CardView>
 
-    <io.github.skincanorg.skincan.widget.RoundedButton
+    <io.github.skincanorg.skincan.widget.button.RoundedButton
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_marginHorizontal="32dp"
+            android:layout_marginHorizontal="54dp"
             android:layout_marginBottom="12dp"
             android:text="@string/scan"
             app:layout_constraintStart_toStartOf="parent"
@@ -37,14 +37,14 @@
             app:layout_constraintBottom_toTopOf="@id/btn_cancel"
             android:id="@+id/btn_scan"/>
 
-    <io.github.skincanorg.skincan.widget.RoundedButton
+    <io.github.skincanorg.skincan.widget.button.RoundedStrokeButton
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:layout_marginBottom="12dp"
+            android:layout_marginBottom="38dp"
             android:text="@android:string/cancel"
             app:layout_constraintStart_toStartOf="@id/btn_scan"
             app:layout_constraintEnd_toEndOf="@id/btn_scan"
-            app:layout_constraintBottom_toBottomOf="@id/image_preview_wrapper"
+            app:layout_constraintBottom_toBottomOf="parent"
             android:id="@+id/btn_cancel"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
